Started porting the tests over to the new system.
