+++ api/opds.py
@@ -37,7 +37,7 @@ from core.util.cdn import cdnify
 from novelist import NoveListAPI
 from lanes import QueryGeneratedLane
 from annotations import AnnotationWriter
-from adobe_vendor_id import AuthdataUtility
+
 
 class CirculationManagerAnnotator(Annotator):
    
@@ -591,6 +591,8 @@ class CirculationManagerAnnotator(Annotator):
 
         def refresh(credential):
             credential.credential = str(uuid.uuid1())
+        # Avoid circular import    
+        from adobe_vendor_id import AuthdataUtility
         patron_identifier = Credential.lookup(
             _db, internal, AuthdataUtility.ADOBE_ACCOUNT_ID_PATRON_IDENTIFIER, patron,
             refresher_method=refresh, allow_persistent_token=True
@@ -639,6 +641,8 @@ class CirculationManagerAnnotator(Annotator):
         cached = self._adobe_id_tags.get(patron_identifier)
         if cached is None:
             cached = []
+            # Avoid circular import    
+            from adobe_vendor_id import AuthdataUtility
             authdata = AuthdataUtility.from_config()
             if authdata:
                 # TODO: We would like to call encode() here, and have
